Description
This problem can be solved using Two pointer. Since our first element is already placed at zero position and it is a unique element, so we quickly run through entire loop to scan unique element. If the current element and next element are the same then we just keep on going till we find different element. Once we find the different element, it is inserted at index 1 because index 0 is taken by the first unique element. Once this is done, the same scanning is done to find out next unique element and this element is to be inserted at index 2. This process continues until we are done with the unique elements.

Time Complexity: O(n)
Space Complexity: O(1)